+++ core/src/test/java/com/facebook/ktfmt/format/GoogleStyleFormatterKtTest.kt
@@ -18,6 +18,8 @@ package com.facebook.ktfmt.format
 
 import com.facebook.ktfmt.testutil.assertFormatted
 import com.facebook.ktfmt.testutil.assertThatFormatting
+import com.facebook.ktfmt.testutil.defaultTestFormattingOptions
+import org.junit.BeforeClass
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.JUnit4
@@ -100,7 +102,7 @@ class GoogleStyleFormatterKtTest {
         |"""
             .trimMargin()
 
-    assertThatFormatting(code).withOptions(Formatter.GOOGLE_FORMAT).isEqualTo(expected)
+    assertThatFormatting(code).isEqualTo(expected)
     // Don't add more tests here
   }
 
@@ -144,7 +146,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -196,7 +197,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -245,7 +245,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -270,7 +269,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -292,7 +290,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -310,7 +307,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -335,7 +331,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -361,7 +356,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -373,7 +367,6 @@ class GoogleStyleFormatterKtTest {
       |  .format(expression)
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -401,7 +394,6 @@ class GoogleStyleFormatterKtTest {
       |    .secondLink()
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -420,7 +412,6 @@ class GoogleStyleFormatterKtTest {
       |  DUMMY
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -480,7 +471,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -498,7 +488,7 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT)
+      )
 
   @Test
   fun `line breaks inside when expressions and conditions`() =
@@ -529,7 +519,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
       )
 
   @Test
@@ -545,7 +534,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
       )
 
   @Test
@@ -565,7 +553,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
       )
 
   @Test
@@ -581,7 +568,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
       )
 
   @Test
@@ -599,7 +585,7 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT)
+      )
 
   @Test
   fun `Arguments are blocks`() =
@@ -628,7 +614,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -649,7 +634,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -668,7 +652,7 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT)
+      )
 
   @Test
   fun `named arguments indent their value expression`() =
@@ -685,7 +669,7 @@ class GoogleStyleFormatterKtTest {
       |  )
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT)
+      )
 
   @Test
   fun `breaking long binary operations`() =
@@ -709,7 +693,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -761,7 +744,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -793,7 +775,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   // TODO: there's a bug here - the last case shouldn't break after 'foo'.
@@ -823,7 +804,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -849,7 +829,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -866,7 +845,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -910,7 +888,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -947,7 +924,7 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
             .trimMargin()
-    assertThatFormatting(code).withOptions(Formatter.GOOGLE_FORMAT).isEqualTo(expected)
+    assertThatFormatting(code).isEqualTo(expected)
   }
 
   @Test
@@ -1034,7 +1011,7 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
             .trimMargin()
-    assertThatFormatting(code).withOptions(Formatter.GOOGLE_FORMAT).isEqualTo(expected)
+    assertThatFormatting(code).isEqualTo(expected)
   }
 
   @Test
@@ -1072,7 +1049,7 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
             .trimMargin()
-    assertThatFormatting(code).withOptions(Formatter.GOOGLE_FORMAT).isEqualTo(expected)
+    assertThatFormatting(code).isEqualTo(expected)
   }
 
   @Test
@@ -1105,7 +1082,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
             .trimMargin(),
-        formattingOptions = Formatter.GOOGLE_FORMAT,
         deduceMaxWidth = false)
   }
 
@@ -1138,7 +1114,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
             .trimMargin(),
-        formattingOptions = Formatter.GOOGLE_FORMAT,
         deduceMaxWidth = false)
   }
 
@@ -1195,7 +1170,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -1219,7 +1193,7 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT)
+      )
 
   @Test
   fun `chained calls that don't fit in one line`() =
@@ -1238,7 +1212,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -1289,7 +1262,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -1307,7 +1279,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -1336,7 +1307,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -1353,7 +1323,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -1384,7 +1353,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -1402,7 +1370,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -1431,7 +1398,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -1448,7 +1414,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -1471,7 +1436,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -1490,7 +1454,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -1505,7 +1468,6 @@ class GoogleStyleFormatterKtTest {
       |)
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -1522,7 +1484,6 @@ class GoogleStyleFormatterKtTest {
       |  ) -> Unit
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -1545,7 +1506,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -1566,7 +1526,7 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT)
+      )
 
   @Test
   fun `array-literal in annotation`() =
@@ -1609,7 +1569,6 @@ class GoogleStyleFormatterKtTest {
       |class Host
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -1673,7 +1632,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -1711,7 +1669,6 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT,
           deduceMaxWidth = true)
 
   @Test
@@ -1731,7 +1688,7 @@ class GoogleStyleFormatterKtTest {
       |}
       |"""
               .trimMargin(),
-          formattingOptions = Formatter.GOOGLE_FORMAT)
+      )
 
   @Test
   fun `trailing commas in enums`() {
@@ -1829,11 +1786,17 @@ class GoogleStyleFormatterKtTest {
         |}
         |"""
             .trimMargin()
-    assertThatFormatting(code).withOptions(Formatter.GOOGLE_FORMAT).isEqualTo(expected)
+    assertThatFormatting(code).isEqualTo(expected)
   }
 
   companion object {
     /** Triple quotes, useful to use within triple-quoted strings. */
     private const val TQ = "\"\"\""
+
+    @JvmStatic
+    @BeforeClass
+    fun setUp(): Unit {
+      defaultTestFormattingOptions = Formatter.GOOGLE_FORMAT
+    }
   }
 }