+++ core/src/test/java/com/facebook/ktfmt/testutil/KtfmtTruth.kt
@@ -26,6 +26,8 @@ import com.google.common.truth.Truth
 import org.intellij.lang.annotations.Language
 import org.junit.Assert
 
+var defaultTestFormattingOptions: FormattingOptions = Formatter.META_FORMAT
+
 /**
  * Verifies the given code passes through formatting, and stays the same at the end
  *
@@ -44,7 +46,7 @@ import org.junit.Assert
  */
 fun assertFormatted(
     @Language("kts") code: String,
-    formattingOptions: FormattingOptions = Formatter.META_FORMAT,
+    formattingOptions: FormattingOptions = defaultTestFormattingOptions,
     deduceMaxWidth: Boolean = false,
 ) {
   val first = code.lines().first()
@@ -81,7 +83,7 @@ fun assertThatFormatting(@Language("kts") code: String): FormattedCodeSubject {
 
 class FormattedCodeSubject(metadata: FailureMetadata, private val code: String) :
     Subject(metadata, code) {
-  private var options: FormattingOptions = Formatter.META_FORMAT
+  private var options: FormattingOptions = defaultTestFormattingOptions
   private var allowTrailingWhitespace = false
 
   fun withOptions(options: FormattingOptions): FormattedCodeSubject {